HT_23/requirements.txt
django==4.1.5
djangorestframework==3.14.0
requests==2.28.2
Celery==5.2.7